requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
pillow
jinja2
